requirements.txt
PyQt5>=5.15.9
configparser
python-telegram-bot>=13.7
fpdf2
qrcode[pil]